Reduce headers dependency
