+++ ydb/core/nbs/cloud/blockstore/libs/storage/partition_direct/erase_request.cpp
@@ -2,6 +2,9 @@
 
 #include <ydb/core/nbs/cloud/storage/core/libs/common/future_helper.h>
 
+#include <ydb/library/actors/core/log.h>
+#include <ydb/library/services/services.pb.h>
+
 namespace NYdb::NBS::NBlockStore::NStorage::NPartitionDirect {
 
 ////////////////////////////////////////////////////////////////////////////////
