+++ ydb/core/nbs/cloud/blockstore/libs/storage/partition_direct/load_actor_adapter.cpp
@@ -7,9 +7,14 @@
 #include <ydb/core/nbs/cloud/storage/core/libs/common/guarded_sglist.h>
 #include <ydb/core/nbs/cloud/storage/core/libs/common/sglist.h>
 
+#include <ydb/core/base/appdata_fwd.h>
+
 #include <ydb/library/actors/core/actorsystem.h>
+#include <ydb/library/actors/core/log.h>
+#include <ydb/library/services/services.pb.h>
 
 using namespace NThreading;
+using namespace NActors;
 
 namespace NYdb::NBS::NBlockStore::NStorage::NPartitionDirect {
 
@@ -20,7 +25,7 @@ TLoadActorAdapter::TLoadActorAdapter(
     : FastPathService(std::move(fastPathService))
 {}
 
-void TLoadActorAdapter::Bootstrap(const NActors::TActorContext& ctx)
+void TLoadActorAdapter::Bootstrap(const TActorContext& ctx)
 {
     Y_UNUSED(ctx);
     Become(&TThis::StateWork);
@@ -33,7 +38,7 @@ void TLoadActorAdapter::Bootstrap(const NActors::TActorContext& ctx)
 
 void TLoadActorAdapter::HandleWriteBlocksRequest(
     const TEvService::TEvWriteBlocksRequest::TPtr& ev,
-    const NActors::TActorContext& ctx)
+    const TActorContext& ctx)
 {
     const auto* msg = ev->Get();
 
@@ -72,7 +77,7 @@ void TLoadActorAdapter::HandleWriteBlocksRequest(
         std::move(request));
 
     future.Subscribe(
-        [actorSystem = NActors::TActivationContext::ActorSystem(),
+        [actorSystem = TActivationContext::ActorSystem(),
          sender = ev->Sender,
          selfId = ctx.SelfID,
          cookie = ev->Cookie,
@@ -93,7 +98,7 @@ void TLoadActorAdapter::HandleWriteBlocksRequest(
 
 void TLoadActorAdapter::HandleReadBlocksRequest(
     const TEvService::TEvReadBlocksRequest::TPtr& ev,
-    const NActors::TActorContext& ctx)
+    const TActorContext& ctx)
 {
     const auto* msg = ev->Get();
 
@@ -116,7 +121,7 @@ void TLoadActorAdapter::HandleReadBlocksRequest(
         request);
 
     future.Subscribe(
-        [actorSystem = NActors::TActivationContext::ActorSystem(),
+        [actorSystem = TActivationContext::ActorSystem(),
          sender = ev->Sender,
          selfId = ctx.SelfID,
          cookie = ev->Cookie,
@@ -151,21 +156,21 @@ void TLoadActorAdapter::HandleReadBlocksRequest(
 STFUNC(TLoadActorAdapter::StateWork)
 {
     LOG_DEBUG(
-        NActors::TActivationContext::AsActorContext(),
+        TActivationContext::AsActorContext(),
         NKikimrServices::NBS_PARTITION,
         "Processing event: %s from sender: %lu",
         ev->GetTypeName().data(),
         ev->Sender.LocalId());
 
     switch (ev->GetTypeRewrite()) {
-        cFunc(NActors::TEvents::TEvPoison::EventType, PassAway);
+        cFunc(TEvents::TEvPoison::EventType, PassAway);
 
         HFunc(TEvService::TEvWriteBlocksRequest, HandleWriteBlocksRequest);
         HFunc(TEvService::TEvReadBlocksRequest, HandleReadBlocksRequest);
 
         default:
             LOG_DEBUG_S(
-                NActors::TActivationContext::AsActorContext(),
+                TActivationContext::AsActorContext(),
                 NKikimrServices::NBS_PARTITION,
                 "Unhandled event type: " << ev->GetTypeRewrite()
                                          << " event: " << ev->ToString());
@@ -176,16 +181,16 @@ STFUNC(TLoadActorAdapter::StateWork)
 ///////////////////////////////////////////////////////////////////////////////
 
 TActorId CreateLoadActorAdapter(
-    const NActors::TActorId& owner,
+    const TActorId& owner,
     std::shared_ptr<TFastPathService> fastPathService)
 {
     auto actor =
         std::make_unique<TLoadActorAdapter>(std::move(fastPathService));
 
-    return NActors::TActivationContext::Register(
+    return TActivationContext::Register(
         actor.release(),
         owner,
-        NActors::TMailboxType::ReadAsFilled,
+        TMailboxType::ReadAsFilled,
         NKikimr::AppData()->SystemPoolId);
 }
