+++ ydb/core/nbs/cloud/blockstore/libs/storage/partition_direct/read_request.h
@@ -2,6 +2,7 @@
 
 #include "vchunk_config.h"
 
+#include <ydb/core/nbs/cloud/blockstore/libs/service/public.h>
 #include <ydb/core/nbs/cloud/blockstore/libs/service/request.h>
 #include <ydb/core/nbs/cloud/blockstore/libs/storage/partition_direct/direct_block_group.h>
 #include <ydb/core/nbs/cloud/blockstore/libs/storage/partition_direct/dirty_map/dirty_map.h>
