+++ ydb/core/nbs/cloud/blockstore/libs/storage/partition_direct/vchunk.h
@@ -1,20 +1,21 @@
 #pragma once
 
-#include "direct_block_group.h"
+#include "public.h"
+
 #include "erase_request.h"
 #include "flush_request.h"
 #include "vchunk_config.h"
 #include "write_request.h"
 
+#include <ydb/core/nbs/cloud/blockstore/config/config.h>
+#include <ydb/core/nbs/cloud/blockstore/libs/common/thread_checker.h>
 #include <ydb/core/nbs/cloud/blockstore/libs/diagnostics/trace_helpers.h>
 #include <ydb/core/nbs/cloud/blockstore/libs/diagnostics/vchunk_counters.h>
-#include <ydb/core/nbs/cloud/blockstore/libs/service/context.h>
 #include <ydb/core/nbs/cloud/blockstore/libs/service/public.h>
 #include <ydb/core/nbs/cloud/blockstore/libs/service/request.h>
 #include <ydb/core/nbs/cloud/blockstore/libs/storage/partition_direct/dirty_map/dirty_map.h>
 
 #include <ydb/core/nbs/cloud/storage/core/libs/common/public.h>
-#include <ydb/core/nbs/cloud/storage/core/libs/coroutine/executor.h>
 
 #include <library/cpp/monlib/dynamic_counters/counters.h>
