+++ ydb/core/nbs/cloud/blockstore/libs/storage/partition_direct/write_request.cpp
@@ -1,7 +1,12 @@
 #include "write_request.h"
 
+#include "direct_block_group.h"
+
 #include <ydb/core/nbs/cloud/blockstore/libs/common/constants.h>
 #include <ydb/core/nbs/cloud/blockstore/libs/diagnostics/trace_helpers.h>
+
+#include <ydb/library/actors/core/log.h>
+#include <ydb/library/services/services.pb.h>
 
 namespace NYdb::NBS::NBlockStore::NStorage::NPartitionDirect {
