+++ ydb/core/nbs/cloud/blockstore/libs/storage/partition_direct/write_request.h
@@ -1,12 +1,17 @@
 #pragma once
 
+#include "public.h"
+
+#include "vchunk_config.h"
+
 #include <ydb/core/nbs/cloud/blockstore/config/config.h>
 #include <ydb/core/nbs/cloud/blockstore/config/protos/storage.pb.h>
 #include <ydb/core/nbs/cloud/blockstore/libs/service/context.h>
 #include <ydb/core/nbs/cloud/blockstore/libs/service/request.h>
-#include <ydb/core/nbs/cloud/blockstore/libs/storage/partition_direct/direct_block_group.h>
 #include <ydb/core/nbs/cloud/blockstore/libs/storage/partition_direct/dirty_map/location.h>
-#include <ydb/core/nbs/cloud/blockstore/libs/storage/partition_direct/vchunk_config.h>
+
+#include <ydb/library/actors/core/actorsystem.h>
+#include <ydb/library/actors/wilson/wilson_span.h>
 
 namespace NYdb::NBS::NBlockStore::NStorage::NPartitionDirect {
