+++ ydb/core/nbs/cloud/blockstore/libs/storage/partition_direct/write_with_pb_replication_request.cpp
@@ -1,6 +1,11 @@
 #include "write_with_pb_replication_request.h"
 
+#include "direct_block_group.h"
+
 #include <ydb/core/nbs/cloud/blockstore/libs/common/constants.h>
+
+#include <ydb/library/actors/core/log.h>
+#include <ydb/library/services/services.pb.h>
 
 namespace NYdb::NBS::NBlockStore::NStorage::NPartitionDirect {
