+++ ydb/core/nbs/cloud/blockstore/libs/storage/partition_direct/ya.make
@@ -4,6 +4,7 @@ GENERATE_ENUM_SERIALIZATION(ddisk_data_copier.h)
 
 SRCS(
     ddisk_data_copier.cpp
+    direct_block_group_impl.cpp
     direct_block_group_in_mem.cpp
     direct_block_group_mock.cpp
     direct_block_group.cpp
@@ -11,11 +12,11 @@ SRCS(
     fast_path_service.cpp
     flush_request.cpp
     load_actor_adapter.cpp
+    part_database.cpp
     part_initschema.cpp
     part_loadstate.cpp
-    part_storevolumeconfig.cpp
     part_storepartitionids.cpp
-    part_database.cpp
+    part_storevolumeconfig.cpp
     partition_direct_actor.cpp
     partition_direct.cpp
     range_translate.cpp
